final/API/requirements.txt
Flask==2.0.3
google-cloud-storage
google-cloud-datastore
protobuf==3.15.6